Youtube Player javascript setup and some html/css setup
Added the "plyrSetup" function to easily add a youtube player onto a div, using an argument that can take an element or querySelector string. Also made the "WebRTCPlayer" not a single line... and changed a little bit of the youtube page css. (Will add more comments explaining YoutubePlayer.js soon)