DROP IN PLAYER PREVIEW
You are now able to drag and drop in the players that are found in the side box.

 It is still a WIP with a slight div boundary bug and a weird Chrome bug that only puts a player down on the second try.

You can also delete players by typing in "delete" in the link input!

This commit is mainly just to show in the meeting and a later commit will probably clean up the CSS and bring the JavaScript "player" class which will hopefully make my code not burn your eyes as much.

Ciao!